v0.2.0-m4

+++ qupath-core/src/main/java/qupath/lib/analysis/stats/Histogram.java
@@ -107,7 +107,7 @@ public class Histogram { // implements Serializable {
 	 * @param ind index of the bin
 	 * @return
 	 */
-	public double getCountsForBin(int ind) {
+	public long getCountsForBin(int ind) {
 		return counts[ind];
 	}
